refactor theme
