+++ packages/ui/src/lib/theme.ts
@@ -1,29 +1,8 @@
-import { createTheme, PaletteColorOptions } from '@mui/material';
+import { createTheme } from '@mui/material';
 
-import { ButtonStyles } from './theme/components/Button.styles';
-import { CssBaselineStyles } from './theme/components/CssBaseline.styles';
-import { InputStyles } from './theme/components/Input.styles';
-import { TypographyStyle } from './theme/components/Typography.styles';
+import { getAllComponents } from './theme/getAllComponents';
 import { palette } from './theme/palette';
 import { typography } from './theme/typography';
-
-declare module '@mui/material' {
-  interface ButtonPropsColorOverrides {
-    accent: true;
-  }
-}
-
-declare module '@mui/material/styles' {
-  interface CustomPalette {
-    accent: PaletteColorOptions;
-  }
-  interface PaletteOptions {
-    accent: PaletteColorOptions;
-  }
-  interface Palette {
-    accent: PaletteColorOptions;
-  }
-}
 
 export const mainTheme = createTheme({
   palette,
@@ -43,9 +22,4 @@ export const mainTheme = createTheme({
   },
 });
 
-mainTheme.components = {
-  ...CssBaselineStyles(mainTheme),
-  ...TypographyStyle(),
-  ...ButtonStyles(mainTheme),
-  ...InputStyles(mainTheme),
-};
+mainTheme.components = getAllComponents(mainTheme);
