--- packages/ui/src/lib/theme/components/CssBaseline.styles.ts
@@ -1,16 +1,0 @@
-import { Components, Theme } from '@mui/material';
-
-export const CssBaselineStyles = (theme: Theme): Partial<Components> => {
-  return {
-    MuiCssBaseline: {
-      styleOverrides: {
-        a: {
-          textDecoration: 'none',
-          color: 'inherit',
-          paddingTop: theme.spacing(2),
-          paddingBottom: theme.spacing(2),
-        },
-      },
-    },
-  };
-};
