--- packages/ui/src/lib/theme/components/Typography.styles.ts
@@ -1,11 +1,0 @@
-import { Components } from '@mui/material';
-
-export const TypographyStyle = (): Partial<Components> => {
-  return {
-    MuiTypography: {
-      defaultProps: {
-        gutterBottom: true,
-      },
-    },
-  };
-};
