+++ packages/ui/src/lib/theme/components/getCssBaselineStyles.ts
@@ -1,0 +1,16 @@
+import { Components, Theme } from '@mui/material';
+
+export const getCssBaselineStyles = (
+  theme: Theme,
+): Pick<Components, 'MuiCssBaseline'> => ({
+  MuiCssBaseline: {
+    styleOverrides: {
+      a: {
+        textDecoration: 'none',
+        color: 'inherit',
+        paddingTop: theme.spacing(2),
+        paddingBottom: theme.spacing(2),
+      },
+    },
+  },
+});
