+++ packages/ui/src/lib/theme/components/getTypographyStyle.ts
@@ -1,0 +1,9 @@
+import { Components } from '@mui/material';
+
+export const getTypographyStyle = (): Pick<Components, 'MuiTypography'> => ({
+  MuiTypography: {
+    defaultProps: {
+      gutterBottom: true,
+    },
+  },
+});
